requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
alembic==1.12.1
celery==5.3.4
redis==5.0.1
stripe==7.8.0
boto3==1.34.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
httpx==0.25.2
pillow==10.1.0
authlib==1.2.1
beautifulsoup4==4.12.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.10.1
isort==5.12.0
flake8==6.1.0

# Database drivers
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Email support
aiosmtplib==3.0.1
email-validator==2.1.0

# Additional utilities
uvicorn[standard]==0.24.0
gunicorn==21.2.0 